requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pillow==10.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
sqlalchemy==2.0.20
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
opencv-python==4.8.0.76
pytesseract==0.3.10
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
